Fix hostid collision in kube backend.
