+++ lib/auth/storage/storage.go
@@ -27,18 +27,23 @@ package storage
 import (
 	"context"
 	"encoding/json"
+	"errors"
+	"io/fs"
 	"strconv"
 	"strings"
 	"time"
 
 	"github.com/coreos/go-semver/semver"
 	"github.com/gravitational/trace"
+	"k8s.io/client-go/util/retry"
 
 	"github.com/gravitational/teleport/api/client/proto"
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/auth/state"
 	"github.com/gravitational/teleport/lib/backend"
+	"github.com/gravitational/teleport/lib/service/servicecfg"
 	"github.com/gravitational/teleport/lib/utils"
+	"github.com/gravitational/teleport/lib/utils/hostid"
 )
 
 const (
@@ -280,4 +285,126 @@ func ReadLocalIdentity(dataDir string, id state.IdentityID) (*state.Identity, er
 	}
 	defer storage.Close()
 	return storage.ReadIdentity(state.IdentityCurrent, id.Role)
+}
+
+// ReadOrGenerateHostID tries to read the `host_uuid` from Kubernetes storage (if available) or local storage.
+// If the read operation returns no `host_uuid`, this function tries to pick it from the first static identity provided.
+// If no static identities were defined for the process, a new id is generated depending on the joining process:
+// - types.JoinMethodEC2: we will use the EC2 NodeID: {accountID}-{nodeID}
+// - Any other valid Joining method: a new UUID is generated.
+// Finally, if a new id is generated, this function writes it into local storage and Kubernetes storage (if available).
+// If kubeBackend is nil, the agent is not running in a Kubernetes Cluster.
+// When facing IsAlreadyExists error, this function will retry reading the host ID from the storage.
+func (p *ProcessStorage) ReadOrGenerateHostID(ctx context.Context, cfg *servicecfg.Config) error {
+	// Read or generate the host ID, which is used to identify the Teleport agent.
+	// If running in Kubernetes, it will read from the Kubernetes secret if available,
+	// otherwise it will read from the local storage.
+	if err := retry.OnError(retry.DefaultRetry, trace.IsAlreadyExists, func() error {
+		return readOrGenerateHostID(ctx, cfg, p.stateStorage)
+	}); err != nil {
+		return trace.Wrap(err)
+	}
+
+	return nil
+}
+
+func readOrGenerateHostID(ctx context.Context, cfg *servicecfg.Config, kubeBackend stateBackend) (err error) {
+	// Load `host_uuid` from different storages. If this process is running in a Kubernetes Cluster,
+	// readHostUUIDFromStorages will try to read the `host_uuid` from the Kubernetes Secret. If the
+	// key is empty or if not running in a Kubernetes Cluster, it will read the
+	// `host_uuid` from local data directory.
+	cfg.HostUUID, err = readHostIDFromStorages(ctx, cfg.DataDir, kubeBackend)
+	if err != nil {
+		if !trace.IsNotFound(err) {
+			if errors.Is(err, fs.ErrPermission) {
+				cfg.Logger.ErrorContext(ctx, "Teleport does not have permission to write to the data directory. Ensure that you are running as a user with appropriate permissions.", "data_dir", cfg.DataDir)
+			}
+			return trace.Wrap(err)
+		}
+		// if there's no host uuid initialized yet, try to read one from the
+		// one of the identities
+		if len(cfg.Identities) != 0 {
+			cfg.HostUUID = cfg.Identities[0].ID.HostUUID
+			cfg.Logger.InfoContext(ctx, "Taking host UUID from first identity.", "host_uuid", cfg.HostUUID)
+		} else {
+			id, err := hostid.Generate(ctx, cfg.JoinMethod)
+			if err != nil {
+				return trace.Wrap(err)
+			}
+			cfg.HostUUID = id
+			cfg.Logger.InfoContext(ctx, "Generating new host UUID", "host_uuid", cfg.HostUUID)
+		}
+		// persistHostUUIDToStorages will persist the host_uuid to the local storage
+		// and to Kubernetes Secret if this process is running on a Kubernetes Cluster.
+		if err := persistHostIDToStorages(ctx, cfg, kubeBackend); err != nil {
+			return trace.Wrap(err)
+		}
+	}
+	return nil
+}
+
+// readHostIDFromStorages tries to read the `host_uuid` value from different storages,
+// depending on where the process is running.
+// If the process is running in a Kubernetes Cluster, this function will attempt
+// to read the `host_uuid` from the Kubernetes Secret. If it does not exist or
+// if it is not running on a Kubernetes cluster the read is done from the local
+// storage: `dataDir/host_uuid`.
+func readHostIDFromStorages(ctx context.Context, dataDir string, kubeBackend stateBackend) (string, error) {
+	if kubeBackend != nil {
+		if hostID, err := loadHostIDFromKubeSecret(ctx, kubeBackend); err == nil && len(hostID) > 0 {
+			return hostID, nil
+		}
+	}
+	// Even if running in Kubernetes fallback to local storage if `host_uuid` was
+	// not found in secret.
+	hostID, err := hostid.ReadFile(dataDir)
+	return hostID, trace.Wrap(err)
+}
+
+// persistHostIDToStorages writes the cfg.HostUUID to local data and to
+// Kubernetes Secret if this process is running on a Kubernetes Cluster.
+func persistHostIDToStorages(ctx context.Context, cfg *servicecfg.Config, kubeBackend stateBackend) error {
+	// Persists the `host_uuid` into Kubernetes Secret for later reusage.
+	// This is required because `host_uuid` is part of the client secret
+	// and Auth connection will fail if we present a different `host_uuid`.
+	if kubeBackend != nil {
+		if err := writeHostIDToKubeSecret(ctx, kubeBackend, cfg.HostUUID); err != nil {
+			// If the storage of the secret fails, don't attempt to write the host id to disk.
+			return trace.Wrap(err)
+		}
+		// Success, write the hostid to disk.
+	}
+
+	if err := hostid.WriteFile(cfg.DataDir, cfg.HostUUID); err != nil {
+		if errors.Is(err, fs.ErrPermission) {
+			cfg.Logger.ErrorContext(ctx, "Teleport does not have permission to write to the data directory. Ensure that you are running as a user with appropriate permissions.", "data_dir", cfg.DataDir)
+		}
+		return trace.Wrap(err)
+	}
+
+	return nil
+}
+
+// loadHostIDFromKubeSecret reads the host_uuid from the Kubernetes secret with
+// the expected key: `/host_uuid`.
+func loadHostIDFromKubeSecret(ctx context.Context, kubeBackend stateBackend) (string, error) {
+	item, err := kubeBackend.Get(ctx, backend.NewKey(hostid.FileName))
+	if err != nil {
+		return "", trace.Wrap(err)
+	}
+	return string(item.Value), nil
+}
+
+// writeHostIDToKubeSecret writes the `host_uuid` into the Kubernetes secret under
+// the key `/host_uuid`.
+func writeHostIDToKubeSecret(ctx context.Context, kubeBackend stateBackend, id string) error {
+	// NOTE: The host uuid should never be updated. If we fail on conflict, we need to backtrack.
+	_, err := kubeBackend.Create(
+		ctx,
+		backend.Item{
+			Key:   backend.NewKey(hostid.FileName),
+			Value: []byte(id),
+		},
+	)
+	return trace.Wrap(err)
 }
