+++ lib/backend/kubernetes/kubernetes_test.go
@@ -48,16 +48,6 @@ func TestBackend_Exists(t *testing.T) {
 		wantErr bool
 	}{
 		{
-			name: "secret does not exist",
-			fields: fields{
-				objects:     nil,
-				namespace:   "test",
-				replicaName: "agent-0",
-			},
-			want:    false,
-			wantErr: false,
-		},
-		{
 			name: "secret exists",
 			fields: fields{
 				objects: []runtime.Object{
@@ -94,6 +84,8 @@ func TestBackend_Exists(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			ctx := context.Background()
+
 			// set namespace env variable
 			if len(tt.fields.namespace) > 0 {
 				t.Setenv(NamespaceEnv, tt.fields.namespace)
@@ -105,14 +97,14 @@ func TestBackend_Exists(t *testing.T) {
 			}
 
 			k8sClient := fake.NewSimpleClientset(tt.fields.objects...)
-			b, err := NewWithClient(k8sClient)
+			b, err := NewWithClient(ctx, k8sClient)
 			if err != nil && !tt.wantErr {
 				require.NoError(t, err)
 			} else if err != nil && tt.wantErr {
 				return
 			}
 
-			require.Equal(t, tt.want, b.Exists(context.TODO()))
+			require.Equal(t, tt.want, b.Exists(ctx))
 		})
 	}
 }
@@ -231,6 +223,8 @@ func TestBackend_Get(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			ctx := context.Background()
+
 			if len(tt.fields.namespace) > 0 {
 				t.Setenv(NamespaceEnv, tt.fields.namespace)
 			}
@@ -240,10 +234,10 @@ func TestBackend_Get(t *testing.T) {
 			}
 
 			k8sClient := fake.NewSimpleClientset(tt.fields.objects...)
-			b, err := NewWithClient(k8sClient)
+			b, err := NewWithClient(ctx, k8sClient)
 			require.NoError(t, err)
 
-			got, err := b.Get(context.TODO(), tt.args.key)
+			got, err := b.Get(ctx, tt.args.key)
 			if (err != nil) && (!trace.IsNotFound(err) || !tt.wantNotFound) {
 				require.NoError(t, err)
 				return
@@ -274,27 +268,6 @@ func TestBackend_Put(t *testing.T) {
 		args   args
 		want   *corev1.Secret
 	}{
-		{
-			name: "secret does not exist and should be created",
-			fields: fields{
-				objects:     nil,
-				namespace:   "test",
-				replicaName: "agent-0",
-			},
-			args: args{
-				item: backend.Item{
-					Key:   backend.NewKey("ids", "kube", "current"),
-					Value: payloadTestData,
-				},
-			},
-			want: newSecret(
-				"agent-0-state",
-				"test",
-				map[string][]byte{
-					backendKeyToSecret(backend.NewKey("ids", "kube", "current")): payloadTestData,
-				},
-			),
-		},
 		{
 			name: "secret exists and has keys",
 			args: args{
@@ -328,6 +301,8 @@ func TestBackend_Put(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			ctx := context.Background()
+
 			// set namespace env var
 			if len(tt.fields.namespace) > 0 {
 				t.Setenv(NamespaceEnv, tt.fields.namespace)
@@ -338,9 +313,12 @@ func TestBackend_Put(t *testing.T) {
 				t.Setenv(teleportReplicaNameEnv, tt.fields.replicaName)
 			}
 
+			for _, o := range tt.fields.objects {
+				o.(*corev1.Secret).ResourceVersion = "1"
+			}
 			k8sClient := fake.NewSimpleClientset(tt.fields.objects...)
 
-			b, err := NewWithClient(k8sClient)
+			b, err := NewWithClient(ctx, k8sClient)
 			require.NoError(t, err)
 
 			// k8s fake client does not support apply operations,
@@ -386,12 +364,13 @@ func TestBackend_Put(t *testing.T) {
 				},
 			)
 			// Put upserts the content in the secret
-			_, err = b.Put(context.TODO(), tt.args.item)
+			_, err = b.Put(ctx, tt.args.item)
 			require.NoError(t, err)
 
 			// get secret loads the kubernetes secret to compare.
-			got, err := b.getSecret(context.TODO())
+			got, err := b.getSecret(ctx)
 			require.NoError(t, err)
+			got.ResourceVersion = tt.want.ResourceVersion // ignore resource version
 			require.Equal(t, tt.want, got)
 		})
 	}
