+++ lib/utils/hostid/hostid.go
@@ -17,13 +17,16 @@
 package hostid
 
 import (
+	"context"
 	"errors"
 	"io/fs"
 	"path/filepath"
 	"strings"
 
+	"github.com/google/uuid"
 	"github.com/gravitational/trace"
 
+	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/utils"
 )
 
@@ -52,4 +55,42 @@ func ReadFile(dataDir string) (string, error) {
 		return "", trace.NotFound("host uuid is empty")
 	}
 	return id, nil
+}
+
+// Generate a new host UUID based on the join method.
+// Uses the EC2 node ID for EC2 instances, and generates a new UUID for all other join methods.
+func Generate(ctx context.Context, joinMethod types.JoinMethod) (string, error) {
+	switch joinMethod {
+	case types.JoinMethodToken,
+		types.JoinMethodUnspecified,
+		types.JoinMethodIAM,
+		types.JoinMethodCircleCI,
+		types.JoinMethodKubernetes,
+		types.JoinMethodGitHub,
+		types.JoinMethodGitLab,
+		types.JoinMethodAzure,
+		types.JoinMethodGCP,
+		types.JoinMethodTPM,
+		types.JoinMethodTerraformCloud:
+		// Checking error instead of the usual uuid.New() in case uuid generation
+		// fails due to not enough randomness. It's been known to happen happen when
+		// Teleport starts very early in the node initialization cycle and /dev/urandom
+		// isn't ready yet.
+		rawID, err := uuid.NewRandom()
+		if err != nil {
+			return "", trace.BadParameter("" +
+				"Teleport failed to generate host UUID. " +
+				"This may happen if randomness source is not fully initialized when the node is starting up. " +
+				"Please try restarting Teleport again.")
+		}
+		return rawID.String(), nil
+	case types.JoinMethodEC2:
+		hostUUID, err := utils.GetEC2NodeID(ctx)
+		if err != nil {
+			return "", trace.Wrap(err)
+		}
+		return hostUUID, nil
+	default:
+		return "", trace.BadParameter("unknown join method %q", joinMethod)
+	}
 }
